Fix useless casts in date.h

+++ include/date/date.h
@@ -5152,7 +5152,7 @@ to_stream(std::basic_ostream<CharT, Traits>& os, const CharT* fmt,
                     tm.tm_wday = static_cast<int>(extract_weekday(os, fds));
                     if (os.fail())
                         return os;
-                    tm.tm_yday = static_cast<int>((ld - local_days(ymd.year()/1/1)).count());
+                    tm.tm_yday = (ld - local_days(ymd.year()/1/1)).count();
                     CharT f[3] = {'%'};
                     auto fe = std::begin(f) + 1;
                     if (modified == CharT{'E'})
@@ -5781,7 +5781,7 @@ to_stream(std::basic_ostream<CharT, Traits>& os, const CharT* fmt,
                         tm.tm_wday = static_cast<int>(extract_weekday(os, fds));
                         if (os.fail())
                             return os;
-                        tm.tm_yday = static_cast<int>((ld - local_days(ymd.year()/1/1)).count());
+                        tm.tm_yday = (ld - local_days(ymd.year()/1/1)).count();
                         facet.put(os, os, os.fill(), &tm, std::begin(f), std::end(f));
                     }
 #endif
@@ -5829,7 +5829,7 @@ to_stream(std::basic_ostream<CharT, Traits>& os, const CharT* fmt,
                         tm.tm_wday = static_cast<int>(extract_weekday(os, fds));
                         if (os.fail())
                             return os;
-                        tm.tm_yday = static_cast<int>((ld - local_days(ymd.year()/1/1)).count());
+                        tm.tm_yday = (ld - local_days(ymd.year()/1/1)).count();
                         facet.put(os, os, os.fill(), &tm, std::begin(f), std::end(f));
                     }
 #endif
@@ -5906,7 +5906,7 @@ to_stream(std::basic_ostream<CharT, Traits>& os, const CharT* fmt,
                         tm.tm_wday = static_cast<int>(extract_weekday(os, fds));
                         if (os.fail())
                             return os;
-                        tm.tm_yday = static_cast<int>((ld - local_days(ymd.year()/1/1)).count());
+                        tm.tm_yday = (ld - local_days(ymd.year()/1/1)).count();
                         facet.put(os, os, os.fill(), &tm, std::begin(f), std::end(f));
                     }
 #endif
@@ -7783,8 +7783,8 @@ from_stream(std::basic_istream<CharT, Traits>& is, const CharT* fmt,
             if (ymd.ok())
             {
                 if (wd == not_a_weekday)
-                    wd = static_cast<int>((weekday(sys_days(ymd)) - Sunday).count());
-                else if (wd != static_cast<int>((weekday(sys_days(ymd)) - Sunday).count()))
+                    wd = (weekday(sys_days(ymd)) - Sunday).count();
+                else if (wd != (weekday(sys_days(ymd)) - Sunday).count())
                     goto broken;
                 if (!computed)
                 {
